:art: refactor: 목업 데이터 분리
- 모든 목업 데이터를 src/app/data/mockData.ts 파일로 이동합니다.
